Add additional check before we create the error state task
